chg: :truck: move the scheduler inside dag/helpers.py
